Move ExceedFilePermissions to diki internal utils
