+++ pkg/internal/utils/utils_test.go
@@ -199,4 +199,23 @@ var _ = Describe("utils", func() {
 			Expect(result).To(Equal(map[string][]utils.FileStats{"test": {fooFileStats}}))
 		})
 	})
+
+	DescribeTable("#ExceedFilePermissions",
+		func(filePermissions, filePermissionsMax string, expectedResult bool, errorMatcher gomegatypes.GomegaMatcher) {
+			result, err := utils.ExceedFilePermissions(filePermissions, filePermissionsMax)
+
+			Expect(result).To(Equal(expectedResult))
+			Expect(err).To(errorMatcher)
+		},
+		Entry("should return false when filePermissions do not exceed filePermissionsMax",
+			"0600", "0644", false, BeNil()),
+		Entry("should return false when filePermissions equal filePermissionsMax",
+			"0644", "0644", false, BeNil()),
+		Entry("should return true when filePermissions exceed filePermissionsMax by user permissions",
+			"0700", "0644", true, BeNil()),
+		Entry("should return true when filePermissions exceed filePermissionsMax by group permissions",
+			"0460", "0644", true, BeNil()),
+		Entry("should return true when filePermissions exceed filePermissionsMax by other permissions",
+			"0406", "0644", true, BeNil()),
+	)
 })
